[#8] Improve usage documentation
Currently we state that a Grafana record is needed as the first
parameter of every function offered by the library. But we don't state
how or where to get such a record from. Document how to obtain such a
record.

Also several return status values, signalling different error
conditions, are not listed in the functions' documentation. So add
them.