[#13] Perform ops atomically, to avoid concurrency issues
Operations inside `with-org` need to be performed atomically, to avoid
concurrency issues. Due to the way Grafana REST API is designed, some
operations dealing with resources attached to an organization (like
dashboards and datasources) need to perform two sequential REST API
calls:

* the first one to switch to the desired organization,
* the second one to actually perform the operation on the resource,
  which doesn't include any reference to the organization, just to the
  resource(s).

The issue with this approach is that different users of the dashboard
manager can execute resource operations in different organizations
concurrently. And depending on the interleaving of the execution
sequence of each thread, it may happen that one of the users ends up
performing the operation on the wrong organization. E.g., the first
user performs the first request. But before the second request is
performed, the thread of the second user is scheduled and it performs
the first request. Now when the first user thread is scheduled again
and performs the second request, it's going to perform the request on
a resource of the second organization. Depending on whether such
resource exists or not, or the user has permissions or not, etc., it
may get different results. But in any case, it won't get the correct
result.

So we need to somehow guarantee that those two REST API request are
executed atomically, without any interleaving among concurrent users'
threads.

After reading quite a bit about concurrency and synchronization
options in Clojure, we think that `locking` is the simplest
option (https://purelyfunctional.tv/guide/clojure-concurrency/#lock)

We've done a few tests, and it works under the following
assumption. As long as we instantiate the Grafana record only once,
and pass the same record to every consumer of the API, we are
good. Otherwise, we get back to where we were: no synchronization at
all.

On the other hand, we think this is a reasonable assumption. Consumers
of the dashboard manager library should have a single Integrant key
for :magnet.dashboard-manager/grafana and then #ig/ref it from every
other Integrant key that is a consumer of the Grafana API. Done this
way, we are sure we are using the same underlying Grafana record
everywhere, and locking works as expected.

On the other hand, if we have more that one key of that type, that
would surely mean that we want to interact with more than one Grafana
server. And that's a different scenario. Still, it would be ok, as
each Grafana server needs to synchronize its own calls independently
from each other.

Closes #13