Name Grafana container we build fo rtesting
As we run tests twice, against legacy and latest Grafana images, we
build a container with respective Grafana images twice. That means
between the tests we must shut down already tested container and build
a fresh new one. Naming the container is a handy option, which gives
us flexibility to reference the specific container whenever we need to
build or kill it.